Further refactor + enhance test coverage
