+++ .cargo/config.toml
@@ -1,0 +1,2 @@
+[env]
+MACOSX_DEPLOYMENT_TARGET = "12.7"
