+++ .github/workflows/pull_request.yml
@@ -1,0 +1,94 @@
+on:
+    pull_request:
+    merge_group:
+    push:
+        branches: [master, main]
+
+jobs:
+    linters:
+        name: Linters
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
+            - name: Cache
+              uses: actions/cache@5a3ec84eff668545956fd18022155c47e93e2684 # v4
+              with:
+                  path: |-
+                    ~/.cargo/bin/
+                    ~/.cargo/git/db/
+                    ~/.cargo/registry/cache/
+                    ~/.cargo/registry/index/
+                    ~/.rustup/downloads/
+                    ~/.rustup/update-hashes/
+                    target/
+                  key: ${{ runner.os }}-cargo-${{ github.ref_name }}-${{ hashFiles('.cargo/config.toml', '**/Cargo.toml') }}
+                  restore-keys: ${{ runner.os }}-cargo-${{ github.ref_name }}-
+            - name: Format check
+              run: make format_check
+            - name: Run Clippy
+              run: make clippy
+
+    tests:
+        name: Test
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
+            - name: Cache
+              uses: actions/cache@5a3ec84eff668545956fd18022155c47e93e2684 # v4
+              with:
+                  path: |-
+                    ~/.cargo/bin/
+                    ~/.cargo/git/db/
+                    ~/.cargo/registry/cache/
+                    ~/.cargo/registry/index/
+                    ~/.rustup/downloads/
+                    ~/.rustup/update-hashes/
+                    target/
+                  key: ${{ runner.os }}-cargo-${{ github.ref_name }}-${{ hashFiles('.cargo/config.toml', '**/Cargo.toml') }}
+                  restore-keys: ${{ runner.os }}-cargo-${{ github.ref_name }}-
+            - name: Run Tests
+              run: make test
+    coverage:
+        name: Coverage
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
+            - name: Cache
+              uses: actions/cache@5a3ec84eff668545956fd18022155c47e93e2684 # v4
+              with:
+                  path: |-
+                    ~/.cargo/bin/
+                    ~/.cargo/git/db/
+                    ~/.cargo/registry/cache/
+                    ~/.cargo/registry/index/
+                    ~/.rustup/downloads/
+                    ~/.rustup/update-hashes/
+                    target/
+                  key: ${{ runner.os }}-cargo-${{ github.ref_name }}-${{ hashFiles('.cargo/config.toml', '**/Cargo.toml') }}
+                  restore-keys: ${{ runner.os }}-cargo-${{ github.ref_name }}-
+            - name: Install tools
+              run: cargo +nightly install cargo-llvm-cov
+            - name: Test Coverage
+              run: make coverage
+            - name: Upload Coverage Report
+              uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
+              with:
+                    name: coverage-report
+                    path: target/coverage/html
+                    if-no-files-found: error
+
+    complete:
+        name: Complete
+        needs: [linters, tests, coverage]
+        if: '!cancelled()'
+        runs-on: ubuntu-latest
+        steps:
+            - name: Success
+              if: needs.linters.result == 'success' && needs.tests.result == 'success' && needs.coverage.result == 'success'
+              run: echo 'Success!'
+            - name: Failure
+              if: needs.linters.result != 'success' || needs.tests.result != 'success' || needs.coverage.result != 'success'
+              run: echo 'Failure!' && exit 1
