+++ Cargo.toml
@@ -25,3 +25,6 @@ default = ["serde"]
 serde = ["dep:serde"]
 serde_test = ["serde", "dep:serde", "dep:serde_json"]
 static = []
+
+[lints.rust]
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(coverage, coverage_nightly)'] }
